FastAPIProject1/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
kafka-python
python-dotenv==1.0.0